Fix bug with addr message parsing
In theory, addr v1 messages need to be differentiated by node version.
Very old nodes do not provide timestamps with address records in their
addr messages, so the parse() function expected a version argument,
which was not provided (bug!). Instead of providing the version, it is
assumed that all nodes' versions are recent enough and addr messages
provide timestamps (which is true in practice).